Added ChatScreen To MainScreen
