+++ app/src/main/java/com/dark/neuroverse/activity/TempActivity.kt
@@ -1,26 +1,11 @@
 package com.dark.neuroverse.activity
 
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.dark.ai_module.workers.ModelManager
-import com.dark.neuroverse.ui.screens.NeuroVChatScreen
+import androidx.compose.material3.ExperimentalMaterial3Api
 import com.dark.neuroverse.ui.theme.NeuroVerseTheme
-import com.mp.ai_core.NativeLib
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import org.json.JSONObject
 
 class TempActivity : ComponentActivity() {
 
@@ -29,9 +14,7 @@ class TempActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            NeuroVerseTheme {
-                NeuroVChatScreen()
-            }
+            NeuroVerseTheme {}
         }
     }
 }
