--- app/src/main/java/com/dark/neuroverse/ui/screens/ai-responce.json
@@ -1,13 +1,0 @@
-{
-  "response" : [
-    {
-      "type": "AI",
-      "content": "Wait User I am Searching Online About it"
-    },
-    {
-      "type": "Tool",
-      "content": "web_search_tool",
-      "callAi": true
-    }
-  ]
-}
